requirements.txt
speechrecognition
google-generativeai
python-dotenv